backend/requirements.txt
SQLAlchemy==1.4.22
pytest==7.0.1
uvicorn==0.17.0
fastapi==0.80.0
alembic==1.7.7
python-dotenv==0.20.0
fastapi-framework
requests==2.27.1
fastapi-pagination==0.10.0
sse-starlette==1.2.1
starlette==0.19.1
python-dateutil==2.8.2
pg8000==1.29.4
pytz==2022.7.1
pytest-asyncio==0.21.1
python-jose==3.3.0
httpx==0.26.0